fix: move IsPowerOfTwoUseLog method to math package
